requirements.txt
# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# CLI
click>=8.1.0
typer>=0.9.0
rich>=13.0.0

# Data
pydantic>=2.0.0
pyyaml>=6.0

# Image handling
Pillow>=10.0.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.65.0

# Browser automation
playwright>=1.40.0
